- Se realiza Migracion de estructuira de Base de datos de FIREBIRD a SQL Server - Se modifica la aplicacion para funcionar con Motor de BD SQL Server

+++ src/Totvs.FlatFileSolution/Totvs.FlatFileGenerator.Business/Models/ShippingProcessDetail.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public int ShippingProcessId { get; set; }
-        public int DocumentTypeId { get; set; }
+        public short DocumentTypeId { get; set; }
         public int DocumentId { get; set; }        
         public string FileName { get; set; }
